+++ src/main/java/net/sf/jabref/gui/importer/FetcherPreviewDialog.java
@@ -25,6 +25,7 @@ import javax.swing.table.TableModel;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.JabRefFrame;
+import net.sf.jabref.gui.importer.fetcher.EntryFetcher;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.logic.importer.OutputPrinter;
 import net.sf.jabref.logic.l10n.Localization;
@@ -301,5 +302,11 @@ public class FetcherPreviewDialog extends JDialog implements OutputPrinter {
     @Override
     public void showMessage(String message) {
         JOptionPane.showMessageDialog(this, message);
+    }
+
+    public void showErrorMessage(EntryFetcher fetcher) {
+        showMessage(Localization.lang("Error while fetching from %0", fetcher.getTitle()) +"\n"+
+                        Localization.lang("Please try again later and/or check your network connection."),
+                Localization.lang("Search %0", fetcher.getTitle()), JOptionPane.ERROR_MESSAGE);
     }
 }