+++ src/main/java/net/sf/jabref/gui/importer/ImportInspectionDialog.java
@@ -69,6 +69,7 @@ import net.sf.jabref.gui.filelist.FileListTableModel;
 import net.sf.jabref.gui.groups.GroupTreeNodeViewModel;
 import net.sf.jabref.gui.groups.UndoableChangeEntriesOfGroup;
 import net.sf.jabref.gui.help.HelpAction;
+import net.sf.jabref.gui.importer.fetcher.EntryFetcher;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.gui.renderer.GeneralRenderer;
 import net.sf.jabref.gui.undo.NamedCompound;
@@ -1470,6 +1471,12 @@ public class ImportInspectionDialog extends JDialog implements ImportInspector, 
     @Override
     public void showMessage(String message) {
         JOptionPane.showMessageDialog(this, message);
+    }
+
+    public void showErrorMessage(EntryFetcher fetcher) {
+        showMessage(Localization.lang("Error while fetching from %0", fetcher.getTitle()) +"\n"+
+                        Localization.lang("Please try again later and/or check your network connection."),
+                Localization.lang("Search %0", fetcher.getTitle()), JOptionPane.ERROR_MESSAGE);
     }
 
     public JabRefFrame getFrame() {